+++ src/main/java/com/berry/oss/dao/service/IBucketInfoDaoService.java
@@ -25,7 +25,7 @@ public interface IBucketInfoDaoService extends IService<BucketInfo> {
      * @param name   名称（非必填）
      * @return
      */
-    List<BucketInfoVo> listBucket(Integer userId, String name);
+    List<BucketInfoVo> listBucket(Long userId, String name);
 
-    List<BucketStatisticsInfoDto> getBucketUseInfo(Integer userId);
+    List<BucketStatisticsInfoDto> getBucketUseInfo(Long userId);
 }
